feat: 新增物流蓝图功能，支持区域配置复制粘贴与渲染优化
- **新增物流蓝图物品**：
  - 添加 BlueprintItem 与 BlueprintData 数据结构，支持区域物流配置的复制与粘贴。
  - 蓝图可存储区域内所有面配置、容器升级及链接关系，粘贴时以新锚点重建。
  - 支持生存模式消耗升级卡，创造模式无消耗。

- **蓝图操作流程**：
  - Shift+右键方块设置锚点，再次点击对角方块复制区域（最大4096方块）。
  - 右键空气清除锚点，Shift+右键空气清空蓝图数据。
  - 粘贴时自动应用原组ID，并重建面配置与链接。

- **客户端区域渲染**：
  - 新增 BlueprintRegionRenderer，在玩家手持蓝图时显示区域框线（起点绿色，终点红色）。
  - 渲染距离基于玩家区块渲染距离动态计算，优化性能。

- **配置与界面优化**：
  - 将传输类型基础堆叠大小配置从固定值改为动态获取，支持配置热重载。
  - 合并配置分类，将缓存、网络、性能参数统一至“性能设置”分类。
  - 优化 LinkWorldRenderer 渲染距离计算，基于玩家渲染距离动态调整。
  - 粒子数量随玩家粒子设置（全部/减少/最少）动态调整。

- **修复与调整**：
  - 修复 LogisticsCalculator 中堆叠倍率计算类型不匹配问题。
  - 修复 TransferUtils 中安全限制计算可能的数据溢出。
  - 更新语言文件，补充蓝图相关提示与配置说明。

+++ src/main/java/com/coobird/staticlogistics/transfer/handler/TransferUtils.java
@@ -37,7 +37,7 @@ public class TransferUtils {
         if (destinations.isEmpty() || limit <= 0) return false;
 
         long maxAllowed = SLConfig.getMaxTransferLimit();
-        int safeLimit = (int) Math.min((long) limit, maxAllowed);
+        int safeLimit = (int) Math.min(limit, maxAllowed);
         if (safeLimit < limit) {
             LOGGER.debug("Transfer limit clamped from {} to {}", limit, safeLimit);
         }
